requirements.txt.txt
pandas
matplotlib
flask
scikit-learn
python-dotenv